pyproject.toml
# SPDX-FileCopyrightText: © 2024 Romain Brault <mail@romainbrault.com>
#
# SPDX-License-Identifier: GPL-3.0-or-later

[build-system]
build-backend = "hatchling.build"
requires = [
    "hatchling",
]

[project]
name = "whiteprints"
description = "A Copier-based cookiecutter for creating Python projects managed by uv."
keywords = [
    "Python",
]
license = "GPL-3.0-or-later"
license-files.paths = [
    "LICENSES/GPL-3.0-or-later.txt",
]
maintainers = [
    { name = "Romain Brault", email = "" },
]
authors = [
    { name = "Romain Brault", email = "" },
]
requires-python = ">=3.9,<4.0"
classifiers = [
    "Development Status :: 1 - Planning",
    "Environment :: Console",
    "Framework :: Pytest",
    "Framework :: Sphinx",
    "Framework :: tox",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Typing :: Typed",
]
dynamic = [
    "readme",
    "version",
]
dependencies = [
    "python-dotenv>=1.0.1",
    "rich-click>=1.8.3",
    "typing-extensions>=4.12.2; python_full_version<'3.12'",
]

scripts.whiteprints = "whiteprints.cli.entrypoint:whiteprints"

[tool.uv]
managed = true
dev-dependencies = [
    "scalene>=1.5.19",
    "pynvim>=0.4.3",
    "beartype>=0.18.5",
    "hypothesis>=6.110.1",
    "pylint>=3.2.6",
    "pytest>=8.3.2",
    "pytest-cov>=5.0.0",
    "pytest-html>=4.1.1",
    "pytest-randomly>=3.15.0",
    "pytest-sugar>=1.0.0",
    "pytest-xdist>=3.6.1",
    "xdoctest>=1.2.0",
]
